Move and refactor pages into their own folder
