+++ src/components/sync-wizard/common/CommonStepsBaseInfo.ts
