+++ src/components/sync-wizard/common/steps/GeneralInfoStep.jsx
