+++ src/components/sync-wizard/data/steps/CategoryOptionsSelectionStep.tsx
