+++ src/components/sync-wizard/data/steps/PeriodSelectionStep.tsx
