+++ src/pages/app/themes/dhis2-legacy.theme.js
