+++ src/utils/testing.js
@@ -8,8 +8,8 @@ import OldMuiThemeProvider from "material-ui/styles/MuiThemeProvider";
 import { MuiThemeProvider } from "@material-ui/core/styles";
 import { SnackbarProvider } from "d2-ui-components";
 
-import { muiTheme } from "../components/app/themes/dhis2.theme";
-import muiThemeLegacy from "../components/app/themes/dhis2-legacy.theme";
+import { muiTheme } from "../pages/app/themes/dhis2.theme";
+import muiThemeLegacy from "../pages/app/themes/dhis2-legacy.theme";
 
 // DHIS2 expects a browser environment, add some required keys to the global node namespace
 Object.assign(global, {
